refactor. GlobeSearch 부분 리팩토링
